refactor(constants): 숫자와 출력 문자열을 구분
- 별도로 constants 디렉토리 생성
- NumberConstant와 OutputMessage로 구분
