+++ src/main/java/baseball/constants/NumberConstant.java
@@ -1,0 +1,9 @@
+package baseball.constants;
+
+public class NumberConstant {
+    public final static int ANSWER_LIST_LENGTH = 3;
+    public final static int RANGE_START_NUM = 1;
+    public final static int RANGE_END_NUM = 9;
+    public final static int RESTART_NUM = 1;
+    public final static int END_GAME_NUM = 2;
+}
