+++ src/main/java/baseball/constants/OutputMessage.java
@@ -1,0 +1,16 @@
+package baseball.constants;
+
+public class OutputMessage {
+    public final static String OUT_OF_VALID_NUM_RANGE_ERROR = "잘못된 숫자를 입력하셨습니다.";
+    public final static String NOT_NUMBER_TYPE_ERROR = "잘못된 값을 입력하셨습니다.";
+    public final static String SAME_NUMBER_ERROR = "동일한 숫자를 중복 입력하셨습니다.";
+
+    public final static String NO_MATCH = "낫싱";
+    public final static String BALL = "볼" + " ";
+    public final static String STRIKE = "스트라이크";
+    public final static String FINISH_GAME_CONDITION = "3스트라이크";
+    public final static String CONGRATULATION_MESSAGE = "3개의 숫자를 모두 맞히셨습니다! 게임 종료";
+
+    public final static String ASK_INPUT_MESSAGE = "숫자를 입력해주세요 : ";
+    public final static String ASK_RESTART_MESSAGE = "게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.";
+}
