+++ src/main/java/baseball/domain/Player.java
@@ -6,12 +6,14 @@ import baseball.utils.Validator;
 import java.util.ArrayList;
 import java.util.List;
 
-import static baseball.Constants.RESTART_NUM;
+import static baseball.constants.NumberConstant.RESTART_NUM;
+import static baseball.constants.OutputMessage.ASK_INPUT_MESSAGE;
+import static baseball.constants.OutputMessage.ASK_RESTART_MESSAGE;
 
 public class Player {
 
     public static List<Integer> getAnswerInput() {
-        System.out.print("숫자를 입력해주세요 : ");
+        System.out.print(ASK_INPUT_MESSAGE);
 
         List<Integer> answer = new ArrayList<>();
 
@@ -27,7 +29,7 @@ public class Player {
     }
 
     public static boolean getGameRestartInput() {
-        System.out.println("게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.");
+        System.out.println(ASK_RESTART_MESSAGE);
 
         String input = Console.readLine();
         Validator.validateRestartInput(input);
